Implement writing JSON in compact mode
In compact mode values are sent as an array, without column name.
Column names and types are sent in property "s" (schema) only when Relation message
is encountered in the replication log.

Consumers will have to keep track of the last schema seen, and use it
when deserializing the message.