Store data copy progress in the source database
If a user wants to use data copy option, they will have to create a database named 'pg_output2json'
in the same db cluster that is as the source of the logical replication.

We will automatically create a table 'data_copy_progress' where status and last json sent
will be kept.

This is to prevent copying data for the same tables every time the listener starts.